+++ NEWS.md
@@ -11,6 +11,7 @@ Fixes:
   * Fix internal bug where efficiency matrices did not have "1" at Release for all fish (did not impact results).
   * Fix bug where arrays with no detections for a group\*release combination would cause skewed estimates.
   * Fix estimation of passed fish when using intra-array estimates.
+  * Fix split CJS's not receiving information about intra-array estimates.
 
 Enhancements:
   * The object `detections` now contains a "Valid" column, similarly to the `movements` objects.